Updating multiple? writer to use an array
Prior to this patch, the values could come in either as an array or
bizarrely as a single value. In the case of the single value, if it
was a singular IO object that would mean it would get read and not
rewound.

This ensures that we don't call #select on an IO object but instead
call select on an Array of possible objects that also respond to
select.